+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
  - Refactor the joystick code
  - Fix example app
+ - Rename points to intersectionPoints for collision detection
 
 ## 1.0.0-rc7
  - Moving device related methods (like `fullScreen`) from `util.dart` to `device.dart`